requirements.txt
Django~=3.2.12
Pillow~=9.0.1
django-admin-sortable2~=1.0.4
